feat(lua): dynamically configure lua lsp and enhance diagnostics
Add `on_init` hook to `sumneko_lua` to dynamically configure settings.
This automatically sets the LuaJIT runtime and `VIMRUNTIME` as a workspace
library when no `.luarc.json` is present, enhancing the out-of-the-box
experience for Lua development in Neovim.

Additionally, add `vim` to global diagnostics to prevent false positives
and disable telemetry.