fix(nvim): disable bufferline and adjust showtabline option
The `bufferline.nvim` plugin has been disabled as it is no longer desired.
The `showtabline` option in Neovim's core options has been set to `0` to never show tabs, aligning with the decision to disable `bufferline.nvim`.
Additionally, a minor adjustment was made to the `snipe.lua` configuration, changing the window title for opened buffers and removing an unused buffer format specifier.